requirements.txt
azure-storage-blob
matplotlib==3.8.0
nltk==3.8.1
numpy
pandas==2.2.2
Pillow==10.4.0
plotly==5.9.0
PyYAML==6.0.1
PyYAML==6.0.1
scikit_learn==1.2.2
seaborn==0.13.2
streamlit==1.33.0
streamlit_authenticator==0.3.2
streamlit_option_menu==0.3.12
wordcloud==1.9.3
openpyxl
